fix: wire up Angular router so /international route navigates correctly
- Extract home page into lazy-loaded HomeComponent (home/)
- Reduce AppComponent to a <router-outlet> shell
- Add path '' route for HomeComponent alongside existing /international route
- Use routerLink on the /international link so it navigates within the SPA
